feat(parser): implement a comprehensive C++ header file parser
- Introduced a streaming token-driven parser in `parser_core.go` to handle C++ header files with O(1) memory complexity.
- Added support for parsing preprocessor directives, including `#define` and other directives in `preprocessor.go`.
- Implemented scope management and helper functions in `scope_helpers.go` to manage current scope and access levels.
- Developed template parsing capabilities in `templates.go` for handling template classes, structs, and functions.
- Added parsing for `typedef` and `using` declarations in `using_typedef.go`.
- Implemented variable declaration parsing in `variables.go`, including support for static and const specifiers.
